project estrutura de dados

+++ Generics_Exercises/armazem/Comida.java
@@ -1,9 +1,9 @@
 public class Comida {
     private String nome;
-    private int calorias;
+    private double calorias;
     private double preco;
 
-    public Comida(String nome, int calorias, double preco) {
+    public Comida(String nome, double calorias, double preco) {
         this.nome = nome;
         this.calorias = calorias;
         this.preco = preco;
@@ -15,10 +15,10 @@ public class Comida {
     public void setNome(String nome) {
         this.nome = nome;
     }
-    public int getCalorias() {
+    public double getCalorias() {
         return calorias;
     }
-    public void setCalorias(int calorias) {
+    public void setCalorias(double calorias) {
         this.calorias = calorias;
     }
     public double getPreco() {
@@ -30,6 +30,6 @@ public class Comida {
 
     @Override
     public String toString() {
-        return String.format("[%s] %d R$ %.6f", nome, calorias, preco);
+        return String.format("[%s] %.6f R$ %.6f", nome, calorias, preco);
     }
 }
